Properly handle path with trailing slash in file matching
Today if a path has a trailing slash, the glob pattern will look like
"/path-to-folder//**" (note the double slash). Glob doesn't work with
double slash actually (it doesn't match anything). As a result, the
permission management for fs_read and fs_write is broken when allowed
or denied path has trailing slash.

The fix is to just manually remove the trailing slash.