Fix bugs in execute_bash permission control
Customer wants to allow some commands and deny all other commands.

To do that, today customer needs to add commands to allow list and use a negative
regex in deny list to exclude other commands. That actually doesn't work.

The first issue is deny list regex we use in Rust doesn't work for such negative
look around. When this happens the deny list doesn't match anything. The first
change is to fallback deny list to "*" when regex compilation fails, instead of
ignoring that failure.

The second issue is customer just want an easy way to deny by default. The 2nd
change is to introduce a boolean flag called "denyByDefault" so anything not in
allowed list is denied by default (instead of "ask" by default)