docs: Update agent-format to reflect "name" field now being required.
The "name" field for an agent has become mandatory since this commit:
https://github.com/aws/amazon-q-developer-cli/commit/dc0974fad9c5da624c77311d536069f082698a94#diff-61df6179f46d679df0c7da318067babd72d265d8958e1b94e3a990de0bfba45fL123-L124

The "/agent generate" command also requires providing a name:
https://github.com/aws/amazon-q-developer-cli/commit/a8afb479f8cbb9390e28560ccc198943316ad2c1#diff-f6d402f6ca61a5601d188759a5d5246ceb57016fa6744f5bbbdec85180ff7a2cR1606

This commit is to update the docs to match current functionality.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.