In non-interactive mode, request alternative method on tool rejection
This is a proof-of-concept proposal to improve the user experience when
running in non-interactive mode without trusting all tools. Existing
behavior is to fail and stop the ensure session when an untrusted tool
call is attempted.

When running in interactive mode, the agent+model are able to find
workarounds in some cases -- perhaps a trusted tool can be used instead.

This change handles denied tool calls by asking the model to try an
alternative approach intead of stopping the session.